Refactor out foods tests
